Move local folders store to `wa-sqlite.ts`
