+++ src/lib/components/sidebar-folder/store.ts
@@ -1,5 +1,4 @@
 import { higherSnippetsStore, localDb } from '$lib/sqlite/global'
-
-import { createLocalFoldersStore } from '$lib/utitlities/persistence';
+import { createLocalFoldersStore } from '$lib/sqlite/wa-sqlite'
 
 export const localFoldersStore = await createLocalFoldersStore(localDb, higherSnippetsStore.migrationStateStore)
